Split model files into entity/dao classes. Entities contain only fields describing the model. Dao classes contain all database queries on the respective model. Added dao classes for views and dynamic tables. Replaced DataHelper2 with SqlUtil file, moved queries and other functions to respective dao classes. Added a migrator class to handle all version 8 -> 9 upgrades. Added migrator call inside of OpenHelper onUpgrade function (version 9) Updated database test files. Current status: new schemas are created onUpgrade from version 8 -> 9; although, still need to handle fresh installs and integrate new function calls
