+++ app/src/main/java/com/fieldbook/tracker/database/models/ObservationVariableValueModel.kt
@@ -1,7 +1,6 @@
 package com.fieldbook.tracker.database.models
 
-import com.fieldbook.tracker.database.*
-import io.swagger.client.model.Study
+import com.fieldbook.tracker.database.Row
 
 /**
  * ObservationVariable Attribute/Value Models are used to store
@@ -13,25 +12,4 @@ data class ObservationVariableValueModel(val map: Row) {
         val observation_variable_attribute_db_id: Int by map      //foreign key to unit attribute
         val observation_variable_attribute_value: String? by map //says Name in excel, but should this be value?, string value saved for this attribute
         val observation_variable_id: Int by map //comp. pk 2
-    companion object {
-        const val PK = "internal_id_observation_variable_value"
-        const val FK = "observation_variable_value_db_id"
-        val tableName = "observation_variable_values"
-        val columnDefs by lazy {
-            mapOf(PK to "INTEGER PRIMARY KEY AUTOINCREMENT",
-                    ObservationVariableAttributeModel.FK to "INT REFERENCES ${ObservationVariableAttributeModel.tableName}(${ObservationVariableAttributeModel.PK}) ON DELETE CASCADE",
-                    "observation_variable_attribute_value" to "TEXT",
-                    ObservationVariableModel.FK to "INT REFERENCES ${ObservationVariableModel.tableName}(${ObservationVariableModel.PK}) ON DELETE CASCADE",
-            )}
-
-        fun getVariableValues(id: Int) = withDatabase { db ->
-
-            return@withDatabase db.query(tableName,
-                    where = "${ObservationVariableModel.FK} = ?",
-                    whereArgs = arrayOf("$id")).toTable()
-
-        }
-
-        fun getAll() = withDatabase { it.query(tableName).toTable() }
-    }
 }
